+++ lib/get_navigation/src/routes/url_strategy/web/web_url.dart
@@ -1,5 +1,6 @@
 import "package:flutter_web_plugins/flutter_web_plugins.dart";
 
+/// Removes the hash portion from the URL by setting the URL strategy to use path-based URLs.
 void removeHash() {
   setUrlStrategy(PathUrlStrategy());
 }